# Optimize field match.