--- a/internal/server/token.go
+++ b/internal/server/token.go
@@ -359,7 +359,7 @@ func (s *Server) parseSearchScanBaseTokens(
 						err = errInvalidArgument(valStr)
 						return
 					}
-					valArr = append(valArr, val)
+					valArr[i] = val
 				}
 				t.whereins = append(t.whereins, whereinT{field, -1, valArr})
 				continue