The following optimizations were made:
* use copy() instead of a loop for copying array values;
* use array of values instead of a map for whereins;
* map the field names to index values once per-query, not per point.

The benchmark shows improvement (had to time 100 calls to have reliable data):
```
 name          old time/op    new time/op    delta
FieldMatch-4    2.69µs ± 3%    1.53µs ± 5%  -43.02%  (p=0.000 n=9+10)
```